+++ changelog.md
@@ -10,6 +10,12 @@ Group Members:
 [MATE #2] Jeremy Lee
 [MATE #3] Michaela Yoon
 
+## 0.0.26 - 2018-04-08
+### Modified by Jeremy
+- controllers/Welcome.php: Added logic to get equipment sets
+- public/csv/Equipment_Sets.csv: changed names so that they match data from Accessries.csv
+- assets/js/welcome.js: equipment set function call.
+
 ## 0.0.25 - 2018-04-07
 ### Modified by Alfred
 - added unit tests for all entity models